+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 0.0.6 [July 25, 2017]
+* Fix edge case in web error tracking
+
+### 0.0.5 [July 25, 2017]
+* Automate Resque integration
+
 ### 0.0.4 [July 24, 2017]
 * Handle exceptions in other middlewares for response size tracking